+++ CHANGELOG.md
@@ -6,6 +6,7 @@
  * all deprecated rn_base classes replaced
  * #59 all classes moves to PSR-4 namespaces
  * BC: class `tx_cfcleaguefe_util_Memento` removed
+ * Fix #9: prepend zero for round in matchtable creation
 
 ## v1.7.1 (13.05.2021)
  * Integrate Github actions